Add ws.go and fix prune mode
