+++ btc/node.go
@@ -1,7 +1,6 @@
 package btc
 
 import (
-	"encoding/json"
 	"net/http"
 	"strconv"
 	"time"
@@ -9,27 +8,9 @@ import (
 	"github.com/SwingbyProtocol/tx-indexer/pubsub"
 	"github.com/ant0ine/go-json-rest/rest"
 	"github.com/gorilla/websocket"
-	uuid "github.com/satori/go.uuid"
 	log "github.com/sirupsen/logrus"
 )
 
-const (
-	WATCHTXS   = "watchTxs"
-	UNWATCHTXS = "unwatchTxs"
-	GETTXS     = "getTxs"
-)
-
-type WsPayloadTx struct {
-	Action  string `json:"action"`
-	Address string `json:"address"`
-	Tx      *Tx    `json:"tx"`
-}
-
-type WsPayloadTxs struct {
-	Action  string `json:"action"`
-	Address string `json:"address"`
-	Tx      []*Tx  `json:"txs"`
-}
 type Node struct {
 	blockchain *BlockChain
 	index      *Index
@@ -72,10 +53,9 @@ func (node *Node) Start() {
 			len(node.storage.txs),
 		)
 		log.Info(node.blockchain.blocktimes)
-		if len(node.index.lists) >= 6 {
-			for _, m := range node.index.lists[:6] {
-				count := node.index.counter[m.Address]
-				log.Infof("  c: %7d %7d addr: %60s txid: %s", m.Time, count, m.Address, m.Txid)
+		if len(node.storage.txscores) >= 6 {
+			for _, m := range node.storage.txscores[:6] {
+				log.Infof(" txid: %s time: %7d", m.Txid, m.Time)
 			}
 		}
 		GetMu().RUnlock()
@@ -98,7 +78,13 @@ func (node *Node) SubscribeTx() {
 func (node *Node) SubscribeBlock() {
 	for {
 		block := <-node.blockchain.waitchan
-		node.index.RemoveIndexWithTxBefore(node.blockchain, node.storage)
+		//node.index.RemoveIndexWithTxBefore(node.blockchain, node.storage)
+		prunetime, err := node.blockchain.GetPruneBlockTime()
+		if err != nil {
+			log.Info(err)
+		}
+		node.storage.RemoveTxsWIthPruneTime(prunetime)
+
 		newTxs := block.UpdateTxs(node.storage)
 		count := 0
 		for _, tx := range newTxs {
@@ -189,114 +175,4 @@ func res500(w rest.ResponseWriter, r *rest.Request) {
 	w.WriteHeader(http.StatusInternalServerError)
 	res := []string{}
 	w.WriteJson(res)
-}
-
-func (node *Node) WsHandler(w http.ResponseWriter, r *http.Request) {
-	node.upgrader.CheckOrigin = func(r *http.Request) bool {
-		return true
-	}
-
-	// Time allowed to read the next pong message from the peer.
-	pongWait := 60 * time.Second
-	// Send pings to peer with this period. Must be less than pongWait.
-	pingPeriod := (pongWait * 9) / 10
-
-	conn, err := node.upgrader.Upgrade(w, r, nil)
-	if err != nil {
-		log.Info("upgrade:", err)
-		return
-	}
-	defer conn.Close()
-	log.Info("WS:Client Connected")
-	client := pubsub.Client{
-		ID:         uuid.Must(uuid.NewV4(), nil).String(),
-		Connection: conn,
-	}
-	client.Connection.SetPongHandler(func(msg string) error {
-		client.Connection.SetReadDeadline(time.Now().Add(pongWait))
-		log.Info("Received PON from client id: ", client.ID)
-		return nil
-	})
-	node.ps.AddClient(client)
-	log.Info("New Client is connected, total: ", len(node.ps.Clients))
-
-	ticker := time.NewTicker(pingPeriod)
-	defer ticker.Stop()
-
-	go func() {
-		for {
-			select {
-			case <-ticker.C:
-				node.ps.PublishPing()
-			}
-		}
-	}()
-
-	for {
-		_, message, err := conn.ReadMessage()
-		if err != nil {
-			log.Info("WS:error:", err)
-			node.ps.RemoveClient(client)
-			break
-		}
-		msg := pubsub.Message{}
-		err = json.Unmarshal(message, &msg)
-		if err != nil {
-			log.Info("This is not correct message payload")
-			break
-		}
-		switch msg.Action {
-		case WATCHTXS:
-			node.ps.Subscribe(&client, msg.Address)
-			log.Infof("new subscriber to Address: -> %s %d %s", msg.Address, len(node.ps.Subscriptions), client.ID)
-			break
-		case UNWATCHTXS:
-			log.Infof("Client want to unsubscribe the Address: -> %s %s", msg.Address, client.ID)
-			node.ps.Unsubscribe(&client, msg.Address)
-			break
-		case GETTXS:
-			log.Infof("Client want to get txs of index Address: -> %s %s", msg.Address, client.ID)
-			resTxs := []*Tx{}
-
-			spentFlag := ""
-			if spentFlag == "send" {
-				spents, err := node.index.GetSpents(msg.Address, node.storage)
-				if err != nil {
-					break
-				}
-				for i := len(spents) - 1; i >= 0; i-- {
-					spents[i].EnableTxSpent(msg.Address, node.storage)
-					resTxs = append(resTxs, spents[i])
-				}
-			} else {
-				txs, err := node.index.GetIns(msg.Address, node.storage)
-				if err != nil {
-					break
-				}
-				for i := len(txs) - 1; i >= 0; i-- {
-					txs[i].EnableTxSpent(msg.Address, node.storage)
-					resTxs = append(resTxs, txs[i])
-				}
-			}
-
-			payload := WsPayloadTxs{"getTxs", msg.Address, resTxs}
-			bytes, err := json.Marshal(payload)
-			if err != nil {
-				log.Info(err)
-			}
-			client.Send(bytes)
-			break
-		default:
-			break
-		}
-	}
-}
-
-func (node *Node) WsPublishMsg(addr string, tx *Tx) {
-	payload := WsPayloadTx{"watchTxs", addr, tx}
-	bytes, err := json.Marshal(payload)
-	if err != nil {
-		log.Info(err)
-	}
-	go node.ps.Publish(addr, bytes, nil)
 }
